Move final jsonStringify test from callSignature.spec.ts (call signature return)
- Moved test from packages/run-types/src/runType/member/callSignature.spec.ts:161-168
- Completed moving all jsonStringify tests from callSignature.spec.ts
- Updated progress tracker: 71 tests moved total
- All tests passing
